Split up the Models
