+++ api/db/database.py
@@ -1,5 +1,6 @@
 from sqlmodel import SQLModel, create_engine
-from .models import *
+
+
 # UIsing SQLite here but can easily use PostgreSQL by changing the url
 sqlite_file_name = "db.sqlite3"
 sqlite_url = f"sqlite:///{sqlite_file_name}"
